extracting switch authenticator link click function
