+++ src/v3/src/transformer/button/transformVerifyWithOtherButton.ts
@@ -10,8 +10,6 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import { IdxRemediation } from '@okta/okta-auth-js';
-
 import { IDX_STEP } from '../../constants';
 import {
   IWidgetContext,
@@ -19,6 +17,7 @@ import {
   TransformStepFnWithOptions,
 } from '../../types';
 import { hasMinAuthenticatorOptions, loc } from '../../util';
+import { updateTransactionWithNextStep } from './updateTransactionWithNextStep';
 
 export const transformVerifyWithOtherButton: TransformStepFnWithOptions = ({
   transaction,
@@ -36,37 +35,6 @@ export const transformVerifyWithOtherButton: TransformStepFnWithOptions = ({
     return formbag;
   }
 
-  const onClick = (widgetContext?: IWidgetContext): unknown => {
-    if (typeof widgetContext === 'undefined') {
-      return;
-    }
-    const { setIdxTransaction, setMessage } = widgetContext;
-    const availableSteps = transaction.availableSteps?.filter(
-      ({ name }) => name !== IDX_STEP.SELECT_AUTHENTICATOR_AUTHENTICATE,
-    ) || [];
-    const verifyWithOtherRem = transaction.neededToProceed.find(
-      ({ name }) => name === IDX_STEP.SELECT_AUTHENTICATOR_AUTHENTICATE,
-    ) || {} as IdxRemediation;
-    const availableRemediations = transaction.neededToProceed.filter(
-      ({ name }) => name !== IDX_STEP.SELECT_AUTHENTICATOR_AUTHENTICATE,
-    );
-
-    setMessage(undefined);
-    setIdxTransaction({
-      ...transaction,
-      messages: [],
-      neededToProceed: [
-        verifyWithOtherRem,
-        ...availableRemediations,
-      ],
-      availableSteps: [
-        selectVerifyStep,
-        ...availableSteps,
-      ],
-      nextStep: selectVerifyStep,
-    });
-  };
-
   const { name: step } = selectVerifyStep;
   const listLink: LinkElement = {
     type: 'Link',
@@ -74,7 +42,12 @@ export const transformVerifyWithOtherButton: TransformStepFnWithOptions = ({
     options: {
       label: loc('oie.verification.switch.authenticator', 'login'),
       step,
-      onClick,
+      onClick: (widgetContext?: IWidgetContext): unknown => {
+        if (typeof widgetContext === 'undefined') {
+          return;
+        }
+        updateTransactionWithNextStep(transaction, selectVerifyStep, widgetContext);
+      },
     },
   };
   formbag.uischema.elements.push(listLink);
